+++ pos_orders_history_reprint/models/model.py
@@ -28,8 +28,6 @@ class PosReceipt(models.Model):
     receipt = fields.Char("Receipt")
     pos_reference = fields.Char("Reference")
     receipt_type = fields.Selection([('xml', 'XML'), ('ticket', 'Ticket')], "Receipt Type")
-    # active or not active receipt
-    status = fields.Boolean('Status', default=True)
 
     def save_xml_receipt(self, name, receipt, receipt_type):
         self.create({
@@ -37,19 +35,3 @@ class PosReceipt(models.Model):
             "receipt": receipt,
             "receipt_type": receipt_type,
         })
-
-
-class PosOrder(models.Model):
-    _inherit = "pos.order"
-    order_receipt_id = fields.Many2one('pos.xml_receipt', 'Order Receipt')  # In fact is one2one
-
-    def _create_account_move_line(self, session=None, move=None):
-        orders = self.filtered(lambda o: not o.account_move or o.state == 'paid')
-        res = super(PosOrder, self)._create_account_move_line(session, move)
-        references = [order.pos_reference for order in orders]
-        receipts = self.env['pos.xml_receipt'].search([('pos_reference', 'in', references), ('status', '=', True)])
-        for r in receipts:
-            r.write({
-                'status': False
-            })
-        return res