+++ pos_orders_history_reprint/static/src/js/screens.js
@@ -55,18 +55,28 @@ odoo.define('pos_orders_history_reprint.screens', function (require) {
             var self = this;
             this._super();
             if (this.pos.config.reprint_orders) {
-                this.$('.actions.oe_hidden').removeClass('oe_hidden');
-                this.$('.button.reprint').unbind('click');
-                this.$('.button.reprint').click(function (e) {
-                    var parent = $(this).parents('.order-line');
-                    var id = parseInt(parent.data('id'));
-                    self.click_reprint_order(id);
-                });
+                this.set_reprint_action();
             }
+        },
+        set_reprint_action: function() {
+            var self = this;
+            this.$('.actions.oe_hidden').removeClass('oe_hidden');
+            this.$('.button.reprint').unbind('click');
+            this.$('.button.reprint').click(function (e) {
+                var parent = $(this).parents('.order-line');
+                var id = parseInt(parent.data('id'));
+                self.click_reprint_order(id);
+            });
         },
         click_reprint_order: function (id) {
             this.gui.show_screen('reprint_receipt', {order_id: id});
         },
+        render_list: function(orders) {
+            this._super(orders);
+            if (this.pos.config.reprint_orders) {
+                this.set_reprint_action();
+            }
+        }
     });
 
     screens.ReprintReceiptScreenWidget = screens.ReceiptScreenWidget.extend({
@@ -87,11 +97,21 @@ odoo.define('pos_orders_history_reprint.screens', function (require) {
             var self = this;
             var order = this.get_order();
             this.$('.pos-sale-ticket').hide();
-            new Model('pos.xml_receipt').call('search_read', [[['pos_reference', '=', order.pos_reference],['receipt_type', '=', 'xml']]]).then(function(r) {
+            var receipt = this.pos.get_receipt_by_order_reference_and_type(order.pos_reference, 'xml');
+            if (receipt) {
                 self.$('.pos-sale-ticket').show();
-                self.reprint_receipt = r;
+                self.reprint_receipt = receipt;
                 self.check_handle_auto_print();
-            });
+            } else {
+                new Model('pos.xml_receipt').call('search_read', [[['pos_reference', '=', order.pos_reference], ['receipt_type', '=', 'xml']]]).then(function (r) {
+                    self.$('.pos-sale-ticket').show();
+                    self.reprint_receipt = r;
+                    self.check_handle_auto_print();
+                }, function() {
+                    self.show_popup = true;
+                    self.click_next();
+                });
+            }
         },
         check_handle_auto_print: function() {
             if (this.should_auto_print() && this.reprint_receipt && this.reprint_receipt.length) {