+++ pos_orders_history_reprint/doc/changelog.rst
@@ -3,6 +3,12 @@
 Updates
 =======
 
+`1.0.4`
+-------
+
+- **Fix:** Error related to the Reprint button displaying after barcode scanning in orders history
+- **Improvement:** Speed loading of XML Receipts
+
 `1.0.3`
 -------